RSC: createRscRequestHandler
